feat(number_of_longest_increasing_subsequences): add brute force solution
Added a brute force solution to find all possible
subsequences and count the longest increasing subsequences.

This solution has a time complexity of O(2^n),space
complexity O(2^n) and is notsuitable for large inputs,
but it serves as a reference for understanding the problem.
Since in worst case for each length we have to iterate
through nc1 + nc2 + ... + ncn, which is 2^n.